% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarDemo.R
\name{plotVarDemo}
\alias{plotVarDemo}
\title{Plot Environmental Data}
\usage{
plotVarDemo(varURI, provUri = "", scientificobjectUri = "", token,
  wsUrl = "www.opensilex.org/openSilexAPI/rest/")
}
\arguments{
\item{varURI}{uri of the variable to plot from the \code{\link{variableList}} function or the web service directly}

\item{provUri}{uri of the provenances to select objects from}

\item{scientificobjectUri}{uri of the scientific objects to select}

\item{token}{a token from \code{\link{getToken}} function}

\item{wsUrl}{url of the webservice}
}
\description{
Demonstration function
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
 aToken <- getToken("guest@opensilex.org","guest")
 token <- aToken$data
 vars <- variableList(token = token)
 vars
 plotVarDemo(vars$value[1], token = token)
}

}
